Moved app module to root
