+++ src/app/root/components/root/index.ts
@@ -1,0 +1,1 @@
+export * from './root.component';
