+++ src/app/root/components/root/root.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-root',
+  templateUrl: 'root.component.html',
+  styleUrls: ['root.component.scss']
+})
+export class RootComponent {
+
+}
