+++ src/app/root/root.routes.ts
@@ -1,6 +1,6 @@
 import {Routes} from '@angular/router';
 
-import {videoGamesRoutes} from './videoGames/videoGames.routes';
+import {videoGamesRoutes} from '../videoGames/videoGames.routes';
 
 export const rootRoutes: Routes = [
   ...videoGamesRoutes
